% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/configuration.R
\name{Configuration File}
\alias{Configuration File}
\title{Configuration File}
\description{
This package uses a configuration file to manage basic operational details. It is written in YAML syntax.
}
\references{
For information on the YAML specification, see \url{http://www.yaml.org/spec/1.2/spec.html}.
}

