% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/configuration.R
\name{LoadConfiguration}
\alias{LoadConfiguration}
\title{Parses configuration file and returns a \code{config} object.}
\usage{
LoadConfiguration(configFile = "config.yaml")
}
\arguments{
\item{configFile}{Character vector describing path to configuration file in YAML format.}
}
\value{
Configuration object used by many functions in the \code{nwisnifie2} package.
}
\description{
Parses configuration file and returns a \code{config} object.
}
\examples{
\dontrun{
# Load Configuration File and Start Database Connection
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie2/config.yaml")
 conn <- StartDBConnection(config)
}
}
\seealso{
For proper formatting of the configuration file, see \code{\link{Configuration File}}.
}

